requirements.txt
flask==3.1.0
flask-cors==5.0.1
anthropic==0.43.0
supabase==2.11.0
twilio==9.4.0
python-dotenv==1.0.1
gunicorn==23.0.0
Pillow==11.1.0
